refactor : ViewType 지정방식 개선 그리고 파일 위치 이동
