+++ app/src/main/java/woowacourse/movie/ui/fragment/movieList/MovieListFragment.kt
@@ -8,11 +8,11 @@ import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.model.AdvState
 import woowacourse.movie.model.MovieState
+import woowacourse.movie.ui.adapter.MovieListAdapter
 import woowacourse.movie.ui.adv.AdvDetailActivity
 import woowacourse.movie.ui.fragment.cinemaBottomSheet.CinemaListBottomSheet
-import woowacourse.movie.ui.main.adapter.MovieListAdapter
-import woowacourse.movie.ui.main.itemModel.AdvItemModel
-import woowacourse.movie.ui.main.itemModel.MovieItemModel
+import woowacourse.movie.ui.itemModel.AdvItemModel
+import woowacourse.movie.ui.itemModel.MovieItemModel
 
 class MovieListFragment : Fragment(R.layout.fragment_movie_list) {
     private lateinit var presenter: MovieListContract.Presenter
