+++ app/src/main/java/woowacourse/movie/ui/itemModel/AdvItemModel.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.ui.itemModel
+
+import woowacourse.movie.model.AdvState
+
+class AdvItemModel(
+    val advState: AdvState
+) : ItemModel {
+    companion object {
+        const val type: Int = 2
+    }
+}
