+++ app/src/main/java/woowacourse/movie/ui/itemModel/CinemaItemModel.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.ui.itemModel
+
+import woowacourse.movie.model.CinemaState
+
+class CinemaItemModel(
+    val cinema: CinemaState
+) : ItemModel {
+    companion object {
+        const val type: Int = 4
+    }
+}
