+++ app/src/main/java/woowacourse/movie/ui/itemModel/MovieItemModel.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.ui.itemModel
+
+import woowacourse.movie.model.MovieState
+
+class MovieItemModel(
+    val movieState: MovieState
+) : ItemModel {
+    companion object {
+        const val type: Int = 1
+    }
+}
