+++ app/src/main/java/woowacourse/movie/ui/itemModel/TicketsItemModel.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.ui.itemModel
+
+import woowacourse.movie.model.TicketsState
+
+class TicketsItemModel(
+    val ticketsState: TicketsState
+) : ItemModel {
+    companion object {
+        const val type: Int = 3
+    }
+}
