--- app/src/main/java/woowacourse/movie/ui/main/itemModel/ItemModel.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.ui.main.itemModel
-
-import woowacourse.movie.ui.main.ViewType
-
-interface ItemModel {
-    val viewType: ViewType
-}
