--- app/src/main/java/woowacourse/movie/ui/main/itemModel/MovieItemModel.kt
@@ -1,10 +1,0 @@
-package woowacourse.movie.ui.main.itemModel
-
-import woowacourse.movie.model.MovieState
-import woowacourse.movie.ui.main.ViewType
-
-class MovieItemModel(
-    val movieState: MovieState
-) : ItemModel {
-    override val viewType: ViewType = ViewType.MOVIE
-}
